pyproject.toml
[project]
name = "demo-fastapi"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "wwfyde", email = "wwfyde@163.com" }
]
dependencies = [
    "surreal_app",
    "fastapi[all]>=0.115.0",
    "pip>=24.2",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.13.2",
    "redis>=5.0.8",
    "celery>=5.4.0",
    "logfire[fastapi,psycopg,redis]>=0.52.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.2.1",
    "tqdm>=4.66.5",
    "openai>=1.61.1",
    "python-dotenv[cli]>=1.0.1",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.uv]
#package = false

[project.scripts]
package2 = "package2:main"
init = "demo_fastapi.init_data:main"

[tool.pdm.scripts]
start = "uvicorn demo_fastapi.main:app"         # 生产环境



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "src/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/demo_fastapi",
    # "src/package2"
]

[dependency-groups]
dev = [
    "ipython>=8.12.3",
    "jupyterlab>=4.3.5",
    "notebook>=7.3.2",
    "ruff-lsp>=0.0.60",
    "ruff>=0.9.4",
    "pytest>=8.3.5",
]

[tool.uv.sources]
#surreal_app = { workspace = true }
surreal_app = { path = "packages/surreal_app" }

[tool.uv.workspace]
members = [
    #    "src/demo_fastapi",
    #    "src/package2",
    #    "examples",
    #    "packages/*"
]
